[lexical][lexical-extension] Chore: Use the dispatchCommand primitive and move formatKeyboardShortcut
registerKeyboardShortcuts now calls the internal dispatchCommand
primitive instead of editor.dispatchCommand, and binds $next with
Function.prototype.bind (only when a $dispatch middleware is present)
instead of allocating an arrow closure per dispatch. Move
formatKeyboardShortcut and FormatKeyboardShortcutOptions out of the
lexical core into @lexical/extension - display formatting for menus
doesn't need to bloat the core.
